updated file structure

+++ include/dictionary.h
@@ -1,11 +1,9 @@
 #ifndef DICTIONARY_H
 #define DICTIONARY_H
-#include <ctype.h>
 #include <curl/curl.h>
-#include <string.h>
 
 int get_dict();
-void get_word(char *word, int word_len);
+void get_word(char *word, size_t word_len);
 int check_word(char *word);
 
 #endif
