+++ include/game.h
@@ -1,11 +1,9 @@
 #ifndef GAME_H
 #define GAME_H
 
-#include <ctype.h>
 #include <ncurses.h>
-#include <signal.h>
 #include <string.h>
 
-int game(char *word, int word_len);
+int game(char *word, size_t word_len);
 
 #endif
